39351 - Criação de endpoint para adicionar dono do curso
