+++ sme.gclass.api.worker.rabbit/Controllers/GsaController.cs
@@ -127,21 +127,5 @@ namespace SME.GoogleClassroom.Worker.Rabbit.Controllers
             var retorno = await useCase.Executar(usuarioId);
             return Ok(retorno);
         }
-
-        /// <summary>
-        /// Atribui dono do curso por email de um usuário cadastrado no GSA.
-        /// </summary>
-        /// <response code="200">A consulta foi realizada com sucesso.</response>
-        /// <response code="500">Ocorreu um erro inesperado durante a consulta.</response>
-        /// <response code="601">Houve uma falha de validação durante a consulta.</response>
-        [HttpGet("cursos/atribuir-dono")]
-        [ProducesResponseType(typeof(bool), StatusCodes.Status200OK)]
-        [ProducesResponseType(typeof(RetornoBaseDto), StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(typeof(RetornoBaseDto), 601)]
-        public async Task<IActionResult> AtribuirDonoCurso(string email, long turmaId, long componenteCurricularId, [FromServices] IAtribuirDonoCursoUseCase useCase)
-        {
-            var retorno = await useCase.Executar(email, turmaId, componenteCurricularId);
-            return Ok(retorno);
-        }
     }
 }
